refactor(structure): Renames app directory to src.
